+++ lib/presentation/in_game_app_bar.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:fs_score_card/l10n/app_localizations.dart';
 import 'package:fs_score_card/presentation/light_dark_control.dart';
 import 'package:fs_score_card/presentation/new_game_control.dart';
 import 'package:fs_score_card/presentation/new_score_card_control.dart';
@@ -15,7 +16,7 @@ class InGameAppBar extends ConsumerWidget implements PreferredSizeWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     return AppBar(
-      title: const Text('Scores'),
+      title: Text(AppLocalizations.of(context)!.scores),
       actionsPadding: const EdgeInsets.only(right: 12),
       actions: const [
         NewScoreCardControl(),